requirements.txt
pyyaml
argparse
typing
opencv-python
pylint
black
black[jupyter]
numpy
matplotlib
mypy
types-PyYAML
pillow
ipykernel
matplotlib
torchensemble
torchinfo